Refactoring - move file providers to file manager widget folder
