+++ js/bundles/modules/file_providers.js
@@ -1,6 +1,8 @@
 import DevExpress from "./core";
+import WebAPIFileProvider from "../../ui/file_manager/file_provider/file_provider.webapi";
+import OneDriveFileProvider from "../../ui/file_manager/file_provider/file_provider.onedrive";
 
 module.exports = DevExpress.FileProviders = DevExpress.FileProviders || {};
 
-DevExpress.FileProviders.WebAPI = require("../../file_provider/file_provider.webapi");
-DevExpress.FileProviders.OneDrive = require("../../file_provider/file_provider.onedrive");
+DevExpress.FileProviders.WebAPI = WebAPIFileProvider;
+DevExpress.FileProviders.OneDrive = OneDriveFileProvider;
