+++ js/ui/file_manager/file_provider/file_provider.ajax.js
@@ -1,5 +1,5 @@
-import ajax from "../core/utils/ajax";
-import { Deferred } from "../core/utils/deferred";
+import ajax from "../../../core/utils/ajax";
+import { Deferred } from "../../../core/utils/deferred";
 import { FileProvider } from "./file_provider";
 import ArrayFileProvider from "./file_provider.array";
