+++ js/ui/file_manager/file_provider/file_provider.array.js
@@ -1,4 +1,4 @@
-import { each } from "../core/utils/iterator";
+import { each } from "../../../core/utils/iterator";
 
 import { FileProvider, FileManagerItem } from "./file_provider";
